tests: specify stringsAsFactors=FALSE for R <4.0 compatibility
Under R 3.6, the recently added "TAD is the same for records with the
same time" test is failing with

  Error: id column is required to be numeric
  Backtrace:
      ▆
   1. └─lastdose::lastdose(data, addl_ties = "obs_first") at test-lastdose.R:385:3
   2.   └─lastdose::lastdose_list(data, ...)

This is because, prior to R 4.0, read.csv converts the USUBJID column
to a factor.  In this case, the column isn't a character vector, so
lastdose_list's conversion of col_id to an index isn't triggered.