ci: unpin testthat for R 3.6.3 build
With the ubuntu-20.04 to ubuntu-22.04 update, the R 3.6.3 job fails to
build testthat:

  ../inst/include/testthat/vendor/catch.h:6490:33: error: size of array ‘altStackMem’ is not an integral constant-expression
   6490 |         static char altStackMem[SIGSTKSZ];
        |                                 ^~~~~~~~
  ../inst/include/testthat/vendor/catch.h:6541:45: error: size of array ‘altStackMem’ is not an integral constant-expression
   6541 |     char FatalConditionHandler::altStackMem[SIGSTKSZ] = {};
        |                                             ^~~~~~~~
  make: *** [/opt/R/3.6.3/lib/R/etc/Makeconf:177: test-runner.o] Error 1
  ERROR: compilation failed for package ‘testthat’

The R 3.6.3 job uses an old version of testthat due to hitting a waldo
error ("object 'compare_proxy' not found"), but, trying now with the
latest version of testthat (3.2.1.1), this failure no longer occurs.

So side step the above error by unpinning testthat.